+++ Board/board.py
@@ -148,3 +148,43 @@ class Tablero:
             casilla.append(ficha)
         else:
             raise ValueError("No se puede mover a una casilla ocupada por 2+ fichas rivales.")
+    
+    def hay_movimientos_posibles(self, color, valores_dado): 
+        if self.obtener_bar(color):
+            return self.__puede_salir_de_barra(color, valores_dado)
+        direccion = self.__obtener_direccion(color)
+        return self.__puede_mover_ficha_en_tablero(color, valores_dado, direccion)
+
+    def __obtener_direccion(self, color): #si ficha se mueve para izquierda o derecha
+        if color == "B":
+            return 1
+        else:
+            return -1
+
+    def __calcular_destino_desde_barra(self, color, valor): # calcula movimiento desde barra
+        if color == 'B':
+            return valor - 1
+        else:
+            return 24 - valor
+        return valor - 1 if color == 'B' else 24 - valor
+
+    def __puede_salir_de_barra(self, color, valores_dado):  # Verifica movimiento desde barra
+        for valor in valores_dado:
+            destino = self.__calcular_destino_desde_barra(color, valor)
+            if 0 <= destino <= 23 and self.__es_movimiento_valido(self.__contenedor__[destino], color):
+                return True
+        return False
+
+    def __puede_mover_ficha_en_tablero(self, color, valores_dado, direccion): # Verifica si alguna ficha puede moverse.
+        for i in range(24):
+            casilla = self.__contenedor__[i]
+            if casilla and casilla[-1] == color:
+                for valor in valores_dado:
+                    destino = i + valor * direccion
+                    if self.__es_movimiento_fuera_de_tablero(color, destino):
+                        if self.puede_sacar_ficha(color):
+                            return True
+                        continue
+                    if 0 <= destino <= 23 and self.__es_movimiento_valido(self.__contenedor__[destino], color):
+                        return True
+        return False